test: extract conn_req mocks to common
This step extracts all mocks from conn_req. Other mocks are planned to
be extracted and unified iteratively.
