+++ tests/unit/common/mocks-rdma_cm.c
@@ -1,0 +1,177 @@
+/*
+ * SPDX-License-Identifier: BSD-3-Clause
+ * Copyright 2020, Intel Corporation
+ */
+
+/*
+ * mocks-rmda_cm.c -- librdmacm mocks
+ */
+
+#include <rdma/rdma_cma.h>
+#include <librpma.h>
+
+#include "conn_req.h"
+#include "cmocka_headers.h"
+#include "mocks-rdma_cm.h"
+
+/* mock control entity */
+int Mock_ctrl_defer_destruction = MOCK_CTRL_NO_DEFER;
+
+/*
+ * rdma_destroy_qp -- rdma_destroy_qp() mock
+ */
+void
+rdma_destroy_qp(struct rdma_cm_id *id)
+{
+	check_expected_ptr(id);
+}
+
+/*
+ * rdma_accept -- rdma_accept() mock
+ */
+int
+rdma_accept(struct rdma_cm_id *id, struct rdma_conn_param *conn_param)
+{
+	check_expected(id);
+	assert_non_null(conn_param);
+	assert_null(conn_param->private_data);
+	assert_int_equal(conn_param->private_data_len, 0);
+	assert_int_equal(conn_param->responder_resources, RDMA_MAX_RESP_RES);
+	assert_int_equal(conn_param->initiator_depth, RDMA_MAX_INIT_DEPTH);
+	assert_int_equal(conn_param->flow_control, 1);
+	assert_int_equal(conn_param->retry_count, 7); /* max 3-bit value */
+	assert_int_equal(conn_param->rnr_retry_count, 7); /* max 3-bit value */
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
+	return 0;
+}
+
+/*
+ * rdma_reject -- rdma_reject() mock
+ */
+int
+rdma_reject(struct rdma_cm_id *id, const void *private_data,
+		uint8_t private_data_len)
+{
+	check_expected_ptr(id);
+	assert_null(private_data);
+	assert_int_equal(private_data_len, 0);
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
+	return 0;
+}
+
+/*
+ * rdma_ack_cm_event -- rdma_ack_cm_event() mock
+ */
+int
+rdma_ack_cm_event(struct rdma_cm_event *event)
+{
+	check_expected_ptr(event);
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
+	return 0;
+}
+
+/*
+ * rdma_create_id -- mock of rdma_create_id
+ */
+int
+rdma_create_id(struct rdma_event_channel *channel,
+		struct rdma_cm_id **id, void *context,
+		enum rdma_port_space ps)
+{
+	assert_non_null(id);
+	assert_null(context);
+	assert_int_equal(ps, RDMA_PS_TCP);
+
+	/* allocate (struct rdma_cm_id *) */
+	*id = mock_type(struct rdma_cm_id *);
+	if (*id == NULL) {
+		errno = mock_type(int);
+		return -1;
+	}
+
+	if (!Mock_ctrl_defer_destruction)
+		expect_value(rdma_destroy_id, id, *id);
+
+	return 0;
+}
+
+/*
+ * rdma_destroy_id -- mock of rdma_destroy_id
+ */
+int
+rdma_destroy_id(struct rdma_cm_id *id)
+{
+	check_expected(id);
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
+	return 0;
+}
+
+/*
+ * rdma_resolve_route -- mock of rdma_resolve_route
+ */
+int
+rdma_resolve_route(struct rdma_cm_id *id, int timeout_ms)
+{
+	check_expected(id);
+	assert_int_equal(timeout_ms, RPMA_DEFAULT_TIMEOUT);
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
+	return 0;
+}
+
+/*
+ * rdma_connect -- rdma_connect() mock
+ */
+int
+rdma_connect(struct rdma_cm_id *id, struct rdma_conn_param *conn_param)
+{
+	check_expected(id);
+	assert_non_null(conn_param);
+	assert_null(conn_param->private_data);
+	assert_int_equal(conn_param->private_data_len, 0);
+	assert_int_equal(conn_param->responder_resources, RDMA_MAX_RESP_RES);
+	assert_int_equal(conn_param->initiator_depth, RDMA_MAX_INIT_DEPTH);
+	assert_int_equal(conn_param->flow_control, 1);
+	assert_int_equal(conn_param->retry_count, 7); /* max 3-bit value */
+	assert_int_equal(conn_param->rnr_retry_count, 7); /* max 3-bit value */
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
+	return 0;
+}
+
+/*
+ * rdma_disconnect -- rdma_disconnect() mock
+ */
+int
+rdma_disconnect(struct rdma_cm_id *id)
+{
+	check_expected_ptr(id);
+
+	errno = mock_type(int);
+	if (errno)
+		return -1;
+
+	return 0;
+}
