+++ tests/unit/common/mocks-stdlib.c
@@ -1,0 +1,29 @@
+/*
+ * SPDX-License-Identifier: BSD-3-Clause
+ * Copyright 2020, Intel Corporation
+ */
+
+/*
+ * mocks-stdlib.c -- stdlib mocks
+ */
+
+#include <errno.h>
+#include <stddef.h>
+
+#include "cmocka_headers.h"
+
+void *__real__test_malloc(size_t size);
+
+/*
+ * __wrap__test_malloc -- malloc() mock
+ */
+void *
+__wrap__test_malloc(size_t size)
+{
+	errno = mock_type(int);
+
+	if (errno)
+		return NULL;
+
+	return __real__test_malloc(size);
+}
