+++ tests/unit/conn_req-test/CMakeLists.txt
@@ -5,10 +5,17 @@
 
 include(../../cmake/ctest_helpers.cmake)
 
-build_test_src(conn_req-test
-	conn_req-test.c
-	${LIBRPMA_SOURCE_DIR}/rpma_err.c
-	${LIBRPMA_SOURCE_DIR}/conn_req.c)
+build_test_src(UNIT NAME conn_req-test SRCS
+       conn_req-test.c
+       ${TEST_UNIT_COMMON_DIR}/mocks-ibverbs.c
+       ${TEST_UNIT_COMMON_DIR}/mocks-rdma_cm.c
+       ${TEST_UNIT_COMMON_DIR}/mocks-rpma-conn.c
+       ${TEST_UNIT_COMMON_DIR}/mocks-rpma-info.c
+       ${TEST_UNIT_COMMON_DIR}/mocks-rpma-peer.c
+       ${TEST_UNIT_COMMON_DIR}/mocks-rpma-private_data.c
+       ${TEST_UNIT_COMMON_DIR}/mocks-stdlib.c
+       ${LIBRPMA_SOURCE_DIR}/conn_req.c
+       ${LIBRPMA_SOURCE_DIR}/rpma_err.c)
 
 target_compile_definitions(conn_req-test PRIVATE TEST_MOCK_ALLOC)
