+++ tests/unit/conn_req-test/conn_req-test.c
@@ -13,422 +13,8 @@
 #include "conn_req.h"
 #include "info.h"
 #include "rpma_err.h"
-
-#define MOCK_IP_ADDRESS		"127.0.0.1"
-#define MOCK_SERVICE		"1234" /* a random port number */
-
-/* random values */
-#define MOCK_VERBS		(struct ibv_context *)0x4E4B
-#define MOCK_CQ			(struct ibv_cq *)0x00C0
-#define MOCK_PEER		(struct rpma_peer *)0xFEEF
-#define MOCK_INFO		(struct rpma_info *)0xE6B2
-#define MOCK_CONN		(struct rpma_peer *)0xC004
-#define MOCK_PRIVATE_DATA	((void *)"Random data")
-#define MOCK_PDATA_LEN		(strlen(MOCK_PRIVATE_DATA) + 1)
-
-#define CM_EVENT_CONNECTION_REQUEST_INIT \
-	{NULL, NULL, RDMA_CM_EVENT_CONNECT_REQUEST, 0, {{0}}}
-
-#define CM_EVENT_CONNECT_ERROR_INIT \
-	{NULL, NULL, RDMA_CM_EVENT_CONNECT_ERROR, 0, {{0}}}
-
-#define MOCK_OK 0
-
-/* mock control entities */
-
-#define MOCK_CTRL_DEFER		1
-#define MOCK_CTRL_NO_DEFER	0
-
-/*
- * Cmocka does not allow call expect_* from setup whereas check_* will be called
- * on teardown. So, function creating an object which is called during setup
- * cannot queue any expect_* regarding the function destroying the object
- * which will be called in the teardown.
- */
-static int Mock_ctrl_defer_destruction = MOCK_CTRL_NO_DEFER;
-
-/* mocks */
-
-/*
- * rpma_conn_new -- rpma_conn_new()  mock
- */
-int
-rpma_conn_new(struct rdma_cm_id *id, struct ibv_cq *cq,
-		struct rpma_conn **conn_ptr)
-{
-	check_expected_ptr(id);
-	assert_ptr_equal(cq, MOCK_CQ);
-
-	assert_non_null(conn_ptr);
-
-	struct rpma_conn *conn = mock_type(struct rpma_conn *);
-	if (!conn) {
-		int result = mock_type(int);
-		if (result == RPMA_E_PROVIDER)
-			Rpma_provider_error = mock_type(int);
-
-		return result;
-	}
-
-	*conn_ptr = conn;
-	return 0;
-}
-
-/*
- * rpma_conn_delete -- rpma_conn_delete()  mock
- */
-int
-rpma_conn_delete(struct rpma_conn **conn_ptr)
-{
-	assert_non_null(conn_ptr);
-
-	struct rpma_conn *conn = *conn_ptr;
-	check_expected_ptr(conn);
-
-	int result = mock_type(int);
-	if (result) {
-		if (result == RPMA_E_PROVIDER)
-			Rpma_provider_error = mock_type(int);
-
-		return result;
-	}
-
-	*conn_ptr = NULL;
-	return 0;
-}
-
-/*
- * rpma_conn_set_private_data -- rpma_conn_set_private_data() mock
- */
-int
-rpma_conn_set_private_data(struct rpma_conn *conn,
-		struct rpma_conn_private_data *pdata)
-{
-	assert_non_null(conn);
-	assert_non_null(pdata);
-	check_expected(conn);
-	check_expected(pdata->ptr);
-	check_expected(pdata->len);
-
-	return mock_type(int);
-}
-
-/*
- * ibv_create_cq -- ibv_create_cq() mock
- */
-struct ibv_cq *
-ibv_create_cq(struct ibv_context *context, int cqe, void *cq_context,
-		struct ibv_comp_channel *channel, int comp_vector)
-{
-	assert_ptr_equal(context, MOCK_VERBS);
-	assert_int_equal(cqe, RPMA_DEFAULT_Q_SIZE);
-	assert_null(channel);
-	assert_int_equal(comp_vector, 0);
-
-	struct ibv_cq *cq = mock_type(struct ibv_cq *);
-	if (!cq) {
-		errno = mock_type(int);
-		return NULL;
-	}
-
-	return cq;
-}
-
-/*
- * ibv_destroy_cq -- ibv_destroy_cq() mock
- */
-int
-ibv_destroy_cq(struct ibv_cq *cq)
-{
-	assert_int_equal(cq, MOCK_CQ);
-
-	return mock_type(int);
-}
-
-/*
- * rpma_peer_create_qp -- rpma_peer_create_qp() mock
- */
-int
-rpma_peer_create_qp(struct rpma_peer *peer, struct rdma_cm_id *id,
-		struct ibv_cq *cq)
-{
-	assert_ptr_equal(peer, MOCK_PEER);
-	check_expected_ptr(id);
-	assert_ptr_equal(cq, MOCK_CQ);
-
-	int result = mock_type(int);
-	if (result == RPMA_E_PROVIDER)
-		Rpma_provider_error = mock_type(int);
-
-	return result;
-}
-
-/*
- * rdma_destroy_qp -- rdma_destroy_qp() mock
- */
-void
-rdma_destroy_qp(struct rdma_cm_id *id)
-{
-	check_expected_ptr(id);
-}
-
-/*
- * rdma_accept -- rdma_accept() mock
- */
-int
-rdma_accept(struct rdma_cm_id *id, struct rdma_conn_param *conn_param)
-{
-	check_expected(id);
-	assert_non_null(conn_param);
-	assert_null(conn_param->private_data);
-	assert_int_equal(conn_param->private_data_len, 0);
-	assert_int_equal(conn_param->responder_resources, RDMA_MAX_RESP_RES);
-	assert_int_equal(conn_param->initiator_depth, RDMA_MAX_INIT_DEPTH);
-	assert_int_equal(conn_param->flow_control, 1);
-	assert_int_equal(conn_param->retry_count, 7); /* max 3-bit value */
-	assert_int_equal(conn_param->rnr_retry_count, 7); /* max 3-bit value */
-
-	errno = mock_type(int);
-	if (errno)
-		return -1;
-
-	return 0;
-}
-
-/*
- * rdma_reject -- rdma_reject() mock
- */
-int
-rdma_reject(struct rdma_cm_id *id, const void *private_data,
-		uint8_t private_data_len)
-{
-	check_expected_ptr(id);
-	assert_null(private_data);
-	assert_int_equal(private_data_len, 0);
-
-	errno = mock_type(int);
-	if (errno)
-		return -1;
-
-	return 0;
-}
-
-/*
- * rdma_ack_cm_event -- rdma_ack_cm_event() mock
- */
-int
-rdma_ack_cm_event(struct rdma_cm_event *event)
-{
-	check_expected_ptr(event);
-
-	errno = mock_type(int);
-	if (errno)
-		return -1;
-
-	return 0;
-}
-
-/*
- * rpma_info_new -- mock of rpma_info_new
- */
-int
-rpma_info_new(const char *addr, const char *service, enum rpma_info_side side,
-		struct rpma_info **info_ptr)
-{
-	assert_string_equal(addr, MOCK_IP_ADDRESS);
-	assert_string_equal(service, MOCK_SERVICE);
-	assert_int_equal(side, RPMA_INFO_ACTIVE);
-
-	*info_ptr = mock_type(struct rpma_info *);
-	if (*info_ptr == NULL) {
-		int result = mock_type(int);
-		assert_int_not_equal(result, 0);
-
-		if (result == RPMA_E_PROVIDER)
-			Rpma_provider_error = mock_type(int);
-
-		return result;
-	}
-
-	return 0;
-}
-
-/*
- * rpma_info_delete -- mock of rpma_info_delete
- */
-int
-rpma_info_delete(struct rpma_info **info_ptr)
-{
-	assert_non_null(info_ptr);
-	assert_int_equal(*info_ptr, MOCK_INFO);
-
-	/* if argument is correct it connot fail */
-	return 0;
-}
-
-/*
- * rpma_info_resolve_addr -- mock of rpma_info_resolve_addr
- */
-int
-rpma_info_resolve_addr(const struct rpma_info *info, struct rdma_cm_id *id)
-{
-	assert_int_equal(info, MOCK_INFO);
-	check_expected(id);
-
-	int ret = mock_type(int);
-	if (ret == RPMA_E_PROVIDER)
-		Rpma_provider_error = mock_type(int);
-
-	if (ret == MOCK_OK)
-		expect_value(rdma_resolve_route, id, id);
-
-	return ret;
-}
-
-/*
- * rdma_create_id -- mock of rdma_create_id
- */
-int
-rdma_create_id(struct rdma_event_channel *channel,
-		struct rdma_cm_id **id, void *context,
-		enum rdma_port_space ps)
-{
-	assert_non_null(id);
-	assert_null(context);
-	assert_int_equal(ps, RDMA_PS_TCP);
-
-	/* allocate (struct rdma_cm_id *) */
-	*id = mock_type(struct rdma_cm_id *);
-	if (*id == NULL) {
-		errno = mock_type(int);
-		return -1;
-	}
-
-	if (!Mock_ctrl_defer_destruction)
-		expect_value(rdma_destroy_id, id, *id);
-
-	return 0;
-}
-
-/*
- * rdma_destroy_id -- mock of rdma_destroy_id
- */
-int
-rdma_destroy_id(struct rdma_cm_id *id)
-{
-	check_expected(id);
-
-	errno = mock_type(int);
-	if (errno)
-		return -1;
-
-	return 0;
-}
-
-/*
- * rdma_resolve_route -- mock of rdma_resolve_route
- */
-int
-rdma_resolve_route(struct rdma_cm_id *id, int timeout_ms)
-{
-	check_expected(id);
-	assert_int_equal(timeout_ms, RPMA_DEFAULT_TIMEOUT);
-
-	errno = mock_type(int);
-	if (errno)
-		return -1;
-
-	return 0;
-}
-
-/*
- * rdma_connect -- rdma_connect() mock
- */
-int
-rdma_connect(struct rdma_cm_id *id, struct rdma_conn_param *conn_param)
-{
-	check_expected(id);
-	assert_non_null(conn_param);
-	assert_null(conn_param->private_data);
-	assert_int_equal(conn_param->private_data_len, 0);
-	assert_int_equal(conn_param->responder_resources, RDMA_MAX_RESP_RES);
-	assert_int_equal(conn_param->initiator_depth, RDMA_MAX_INIT_DEPTH);
-	assert_int_equal(conn_param->flow_control, 1);
-	assert_int_equal(conn_param->retry_count, 7); /* max 3-bit value */
-	assert_int_equal(conn_param->rnr_retry_count, 7); /* max 3-bit value */
-
-	errno = mock_type(int);
-	if (errno)
-		return -1;
-
-	return 0;
-}
-
-/*
- * rdma_disconnect -- rdma_disconnect() mock
- */
-int
-rdma_disconnect(struct rdma_cm_id *id)
-{
-	check_expected_ptr(id);
-
-	errno = mock_type(int);
-	if (errno)
-		return -1;
-
-	return 0;
-}
-
-/*
- * rpma_private_data_store -- rpma_private_data_store() mock
- */
-int
-rpma_private_data_store(struct rdma_cm_event *edata,
-		struct rpma_conn_private_data *pdata)
-{
-	const LargestIntegralType allowed_events[] = {
-			RDMA_CM_EVENT_CONNECT_REQUEST,
-			RDMA_CM_EVENT_ESTABLISHED};
-	assert_non_null(edata);
-	assert_in_set(edata->event, allowed_events,
-		sizeof(allowed_events) / sizeof(allowed_events[0]));
-	assert_non_null(pdata);
-	assert_null(pdata->ptr);
-	assert_int_equal(pdata->len, 0);
-
-	pdata->ptr = mock_type(void *);
-	if (pdata->ptr == NULL)
-		return RPMA_E_NOMEM;
-
-	pdata->len = strlen(pdata->ptr) + 1;
-
-	return 0;
-}
-
-/*
- * rpma_private_data_discard -- rpma_private_data_discard() mock
- */
-void
-rpma_private_data_discard(struct rpma_conn_private_data *pdata)
-{
-	assert_non_null(pdata);
-}
-
-void *__real__test_malloc(size_t size);
-
-/*
- * __wrap__test_malloc -- malloc() mock
- */
-void *
-__wrap__test_malloc(size_t size)
-{
-	errno = mock_type(int);
-
-	if (errno)
-		return NULL;
-
-	return __real__test_malloc(size);
-}
+#include "test-common.h"
+#include "mocks-rdma_cm.h"
 
 /* tests */
