+++ tests/unit/info-test/CMakeLists.txt
@@ -5,7 +5,7 @@
 
 include(../../cmake/ctest_helpers.cmake)
 
-build_test_src(info-test
+build_test_src(UNIT NAME info-test SRCS
 	info-test.c
 	${LIBRPMA_SOURCE_DIR}/rpma_err.c
 	${LIBRPMA_SOURCE_DIR}/info.c)
