+++ tests/unit/private_data-test/CMakeLists.txt
@@ -5,7 +5,7 @@
 
 include(../../cmake/ctest_helpers.cmake)
 
-build_test_src(private_data-test
+build_test_src(UNIT NAME private_data-test SRCS
        private_data-test.c
        ${LIBRPMA_SOURCE_DIR}/private_data.c)
